+++ framework/wazuh/__init__.py
@@ -17,7 +17,7 @@ Wazuh is a python package to manage OSSEC.
 
 """
 
-__version__ = '3.1.0'
+__version__ = '3.2.0-alpha1'
 
 
 msg = "\n\nPython 2.7 or newer not found."