+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 
 - [Fixed oscap.py to support new versions of OpenSCAP scanner.](https://github.com/wazuh/wazuh/pull/331)
+- [Fixed bug in cluster when receive an error response from client.](https://github.com/wazuh/wazuh/pull/346)
 
 ## [v3.1.0]
 
@@ -20,7 +21,7 @@ All notable changes to this project will be documented in this file.
 - Added alert and archive output files rotation capabilities.
 - Added rule option to discard field "firedtimes".
 - Added VULS integration for running vulnerability assessments.
-- CIS-CAT Wazuh Module to scan CIS policies. 
+- CIS-CAT Wazuh Module to scan CIS policies.
 
 ### Changed