+++ CHANGELOG.md
@@ -3,10 +3,15 @@ All notable changes to this project will be documented in this file.
 
 ## [v3.2.0]
 
+### Added
+- [Added support for multiple masters in cluster.](https://github.com/wazuh/wazuh/pull/353)
+- [Added type of node in API call `GET/cluster/nodes` and `cluster_control --nodes`](https://github.com/wazuh/wazuh/pull/353)
+
 ### Fixed
 
 - [Fixed oscap.py to support new versions of OpenSCAP scanner.](https://github.com/wazuh/wazuh/pull/331)
 - [Fixed bug in cluster when receive an error response from client.](https://github.com/wazuh/wazuh/pull/346)
+- [Improved cluster logs](https://github.com/wazuh/wazuh/pull/353)
 
 ## [v3.1.0]